[Fixes  #6929] Do not overwrite existing thumbnails if the image is blank
