+++ geonode/base/models.py
@@ -46,7 +46,6 @@ from django.core.files.storage import default_storage as storage
 from mptt.models import MPTTModel, TreeForeignKey
 
 from PIL import Image
-from io import BytesIO
 from resizeimage import resizeimage
 
 from imagekit.models import ImageSpecField
@@ -71,6 +70,7 @@ from geonode.base.enumerations import (
     DEFAULT_SUPPLEMENTAL_INFORMATION)
 from geonode.base.bbox_utils import BBOXHelper
 from geonode.utils import (
+    is_monochromatic_image,
     add_url_params,
     bbox_to_wkt)
 from geonode.groups.models import GroupProfile
@@ -83,6 +83,7 @@ from geonode.notifications_helper import (
 from geonode.people.enumerations import ROLE_VALUES
 from geonode.base.thumb_utils import (
     thumb_path,
+    thumb_size,
     remove_thumbs)
 
 from pyproj import transform, Proj
@@ -1459,14 +1460,16 @@ class ResourceBase(PolymorphicModel, PermissionLevelMixin, ItemBase):
 
         try:
             # Check that the image is valid
-            content_data = BytesIO(image)
-            im = Image.open(content_data)
-            im.verify()  # verify that it is, in fact an image
-
-            name, ext = os.path.splitext(filename)
-            remove_thumbs(name)
+            if is_monochromatic_image(None, image):
+                if not self.thumbnail_url:
+                    raise Exception("Generated thumbnail image is blank")
+                else:
+                    # Skip Image creation
+                    image = None
 
             if upload_path and image:
+                name, ext = os.path.splitext(filename)
+                remove_thumbs(name)
                 actual_name = storage.save(upload_path, ContentFile(image))
                 url = storage.url(actual_name)
                 _url = urlparse(url)
@@ -1504,11 +1507,12 @@ class ResourceBase(PolymorphicModel, PermissionLevelMixin, ItemBase):
                     site_url = settings.SITEURL.rstrip('/') if settings.SITEURL.startswith('http') else settings.SITEURL
                     url = urljoin(site_url, url)
 
+                if thumb_size(name) == 0:
+                    raise Exception("Generated thumbnail image is zero size")
+
                 # should only have one 'Thumbnail' link
-                _links = Link.objects.filter(resource=self, name='Thumbnail')
-                if _links and _links.count() > 1:
-                    _links.delete()
-                obj, created = Link.objects.get_or_create(
+                Link.objects.filter(resource=self, name='Thumbnail').delete()
+                obj = Link.objects.create(
                     resource=self,
                     name='Thumbnail',
                     defaults=dict(
@@ -1532,7 +1536,7 @@ class ResourceBase(PolymorphicModel, PermissionLevelMixin, ItemBase):
             try:
                 Link.objects.filter(resource=self, name='Thumbnail').delete()
                 _thumbnail_url = staticfiles.static(settings.MISSING_THUMBNAIL)
-                obj, created = Link.objects.get_or_create(
+                obj = Link.objects.create(
                     resource=self,
                     name='Thumbnail',
                     defaults=dict(
